Split network requests into server fetch + client UI
Add NetworkRequestsClient (client-side) component to handle UI, tabs, actions, dialogs and axios-based status updates. Convert the page to a server component that fetches network creation requests with getServerSession and passes them as initialRequests to the client. Rename API route files from /api/network/requests to /api/devices/network-creation-requests and adjust imports. Remove related helper functions and React Query hooks for fetching/updating network requests from core/apis/networks and useNetworks (getNetworkRequestsApi, updateNetworkRequestStatusApi, createNetworkApi and useNetworkRequests/useUpdateNetworkRequestStatus) and update submitNetworkRequestApi URL usage. Add server-side logging and fetch caching/revalidation for the requests endpoint.
