+++ src/vertex/app/api/devices/network-creation-requests/[id]/[action]/route.ts
